src/main/java/com/workouttracker/trackanytime/service/UserService.java
package com.workouttracker.trackanytime.service;

import com.workouttracker.trackanytime.model.User;
import com.workouttracker.trackanytime.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

@Service
public class UserService {
    @Autowired
    private UserRepository userRepository;
    @Autowired
    private PasswordEncoder passwordEncoder;

    public User saveUser(User user){
        user.setPassword(passwordEncoder.encode(user.getPassword()));
        return userRepository.save(user);
    }

    public boolean existByEmail(String email){
        return userRepository.findByEmail(email).isPresent();
    }
}
